Batch shared logic (2)
